requirements.txt
bs4
guardrails-ai
ipykernel
ipywidgets
numpy
matplotlib
openai
presidio_analyzer
presidio_anonymizer
sentence_transformers
torch
guardrails-ai[hf-hub]
langchain-core
numpy
chromadb
langchain
pypdf
sentence-transformers
pysqlite3-binary